dht: CID 913541: bucket last-reply time is a time_t
struct bucket stored the 'time of last reply in this bucket' in an int
while everything else in the DHT code uses time_t for absolute times,
so the ctx->now assignment needed a truncating cast to int: a 64-bit
time_t silently wraps there in 2038, and bucket maintenance would then
see every bucket as 600s-stale forever.  Widen the member to time_t
and drop the cast; the only other consumer besides comparisons is a
log line that prints an age difference.

+++ lib/misc/dht/dht-backend-bucket.c
@@ -330,7 +330,7 @@ maybe_new_node(struct lws_dht_ctx *ctx, const lws_dht_hash_t *id,
 		   (nb == NULL || id_cmp(ctx->myid, nb->first) < 0);
 
 	if (confirm == 2)
-		b->time = (int)ctx->now;
+		b->time = ctx->now;
 
 	lws_start_foreach_dll(struct lws_dll2 *, d, lws_dll2_get_head(&b->nodes)) {
 		n = lws_container_of(d, struct node, list);
